requirements.txt
# TraceFlow Lite - Streamlit Cloud Dependencies
pydantic>=2.7
pydantic-settings>=2.3
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
streamlit>=1.36
chromadb>=0.5.0
sentence-transformers>=3.0.0
tenacity>=8.2.3
cachetools>=5.3.3
python-dotenv>=1.0.1
opentelemetry-sdk>=1.39.0
openai>=2.14.0
anthropic>=0.18.0
tiktoken>=0.7.0
